requirements/base.txt
black==21.7b0  # https://github.com/psf/black
celery==5.0.1
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.4.0
django-extra-fields==3.0.0
django-filter==2.3.0
django-redis==4.12.1
djangorestframework==3.12.4
drf-yasg==1.17.1
factory-boy==2.12.0
flake8==3.8.3
ipython==7.17.0
jedi==0.17.2
markdown==3.2.2
python-decouple==3.3
pytest-cov==2.10.0
pytest-django==3.9.0
pytest-mock==3.3.
pytest-timeout==1.4.2
redis==3.5.3
urlwait==1.0
djangorestframework-simplejwt==4.6.0
django-model-utils==4.1.1  # https://github.com/jazzband/django-model-utils
pyjwt==2.1.0
uvicorn[standard]==0.14.0  # https://github.com/encode/uvicorn
django-crispy-forms==1.12.0  # https://github.com/django-crispy-forms/django-crispy-forms
random-password-generator
Pillow==9.0.1
gunicorn